General: Refactor views to be split up into separate files.
